Introduce "assemblies" and deprecate "assembly" in manifests
The "assembly" boolean field was used to decide if a module should be
included in the open-source assembly descriptor. The enterprise assembly
descriptor included a hard coded list of plugins and wasn't generated like
the open-source descriptor.

We need to be able to generate the enterprise assembly descriptor as well
because we now have different artifacts in 2.4 and 3.0, for example.

To do this, we introduced the "assemblies" field which contains a list of
assembly descriptor IDs. The open-souce and enterprise assembly descriptor
templates now get a list of modules for a given assembly ID.

    // graylog-assembly.xml
    {{ range $assembly := index .Assemblies "graylog" }}
      // ...
    {{ end }}

    // graylog-enterprise-assembly.xml
    {{ range $assembly := index .Assemblies "graylog-enterprise" }}
      // ...
    {{ end }}

This makes it possible to configure one or more assemblies for each
module in the manifest. It also allows us to include each module in more
than one assembly.

The CLI tool will emit a warning for manifests which still use the old
"assembly" field.