Allow overriding name of a repository
By specifying a "name" in the manifest, the repository
will be checked out under the given name instead of
auto-detecting the name from the repository name.

This is helpful when a fork of the graylog2-server
repository is used. The Graylog project build relies
on the server repository being checked out as
"graylog2-server".

Without this change, a repository
"Graylog2/graylog2-server-fork.git" would be checked
out as "graylog2-server-fork".

With this change, it can be named "graylog2-server" with
the git remote still pointing at
"Graylog2/graylog2-server-fork.git".